+++ server/methods/_all.js
@@ -3,7 +3,7 @@ import './articles.js';
 import './categories.js';
 import './checkInputs.js';
 import './comments.js';
-import './design.js';
+import './design/design.js';
 import './generic.js';
 import './search.js';
 import './user.js';
